1.2.Introduction_to_API_with_authentication/requirements.txt
requests==2.31.0
python-dotenv==1.0.1